Implement panel reloading
This commit contains a working version of panel reloading. The central
logic is asynchronously loading the next random panels and setting them in
the href of the reload button. That way the user doesn't have to wait
after clicking the button, because the data has already been fetched
from the server.