Fix lock icons and move them to CSS
Previously the lock icons were always open, even when the panel was
locked. This commit fixes that bug.

This commit also refactors the lock icons to be displayed via the background
property in CSS instead of as `img` tags. Also, they are small enough
that it makes sense to inline them in the CSS.