Upgrade css-loader
Update the css-loader module. This also allows us to change the
image loading strategy to "asset", which will automatically choose
between embedding a base64-encoded version of the images in the
CSS or emitting a file, based on the file size. In our case, all
images are small enough to be embedded in the CSS file.